+++ ft_type_s_auxiliary.c
@@ -16,15 +16,16 @@ size_t	ft_work_with_zero_value(t_pmts *pmts, size_t len)
 {
 	if (pmts->prec && !pmts->prec_value)
 		return ((size_t)pmts->zero_value);
-	if (pmts->zero_value < pmts->prec_value && pmts->prec_value <= len)
+	if ((int)pmts->zero_value < pmts->prec_value
+			&& pmts->prec_value <= (int)len)
 		return ((size_t)pmts->prec_value);
 	if (pmts->zero_value < len && !pmts->prec_value)
 		return (len);
-	if (pmts->zero_value < len && pmts->prec_value < len)
+	if (pmts->zero_value < len && pmts->prec_value < (int)len)
 		return ((size_t)pmts->zero_value);
-	if (pmts->zero_value < len && pmts->prec_value > len)
+	if (pmts->zero_value < len && (size_t)pmts->prec_value > len)
 		return (len);
-	if (pmts->zero_value && pmts->prec_value > len)
+	if (pmts->zero_value && (size_t)pmts->prec_value > len)
 		return ((size_t)pmts->zero_value);
 	if (pmts->zero_value && pmts->prec_value)
 		return ((size_t)pmts->zero_value);
@@ -35,26 +36,26 @@ size_t	ft_work_with_value(t_pmts *pmts, size_t len)
 {
 	if (pmts->prec && !pmts->prec_value)
 		return ((size_t)pmts->value);
-	if (pmts->value < pmts->prec_value && pmts->prec_value <= len)
+	if ((int)pmts->value < pmts->prec_value && pmts->prec_value <= (int)len)
 		return ((size_t)pmts->prec_value);
 	if (pmts->value < len && !pmts->prec_value)
 		return (len);
-	if (pmts->value < len && pmts->prec_value < len)
+	if (pmts->value < len && pmts->prec_value < (int)len)
 		return ((size_t)pmts->value);
-	if (pmts->value < len && pmts->prec_value > len)
+	if (pmts->value < len && (size_t)pmts->prec_value > len)
 		return (len);
-	if (pmts->value && pmts->prec_value > len)
-		return ((size_t)pmts->value);
+	if (pmts->value && pmts->prec_value > (int)len)
+		return (pmts->value);
 	if (pmts->value && pmts->prec_value)
-		return ((size_t)pmts->value);
-	return ((size_t)pmts->value);
+		return (pmts->value);
+	return (pmts->value);
 }
 
 size_t	ft_getting_total_len_s(t_pmts *pmts, size_t len)
 {
 	if (pmts->prec_value < 0)
 		return ((size_t)pmts->prec_value * -1);
-	if (pmts->prec_value > len)
+	if ((size_t)pmts->prec_value > len)
 		pmts->prec_value = (int)len;
 	if (pmts->prec && !pmts->prec_value && !pmts->zero && !pmts->value)
 		return (0);
@@ -66,7 +67,7 @@ size_t	ft_getting_total_len_s(t_pmts *pmts, size_t len)
 		return (ft_work_with_zero_value(pmts, len));
 	if (pmts->value)
 		return (ft_work_with_value(pmts, len));
-	if (pmts->prec_value < len && pmts->value > len)
+	if (pmts->prec_value < (int)len && pmts->value > len)
 		return ((size_t)pmts->prec_value);
 	else
 		return (len);
